Vcd multiaccounting and analytical tables. (#67)
* Execution profiles for selenium jobs.

* Remove password value from source control.

* Define account ids for multiple profiles.

* Fixed logging formatting issue.

* Added entry points for vcd data syncing.

* Sync analytical tables with normal tables for vcd data.

* Added coments to SqlScriptsExecutor service.

* Leave empty db connection string.

* Added scripts for sync all accounts to analytic tables. Added ordered revenue to vcd analytic tables.

+++ 3/CakeExtractor.SeleniumApplication/CakeExtractor.SeleniumApplication/SQL/VCD/SyncVcdAnalyticTable.sql
@@ -1,10 +1,11 @@
-﻿--DELETE FROM td.VcdAnalytic
-TRUNCATE TABLE td.VcdAnalytic 
+﻿DECLARE @accId INT = @@param_0;                -- Account ID
+
+DELETE FROM td.VcdAnalytic  WHERE AccountId= @accId;
 
 INSERT INTO td.VcdAnalytic (Date, BrandName, CategoryName, SubcategoryName,ReleaseDate, 
 							Asin, ParentProductAsin, Name,AccountId, Ean, Upc, ApparelSize, ApparelSizeWidth,
 							Binding, Color, ModelStyleNumber,
-							 ShippedRevenue,ShippedUnits, OrderedUnits, ShippedCOGS, FreeReplacements, CustomerReturns)
+							 ShippedRevenue,ShippedUnits, OrderedUnits, ShippedCOGS, FreeReplacements, CustomerReturns, OrderedRevenue)
 select 
 	   summaryMetrics.Date                    as 'Date - PRODUCT',
 	   brand.Name                             as 'Brand',
@@ -28,7 +29,8 @@ select
 	   summaryMetrics.[53] as 'Ordered Units',
 	   summaryMetrics.[55] as 'Shipped COGS',
 	   summaryMetrics.[56] as 'Free Replacements',
-	   summaryMetrics.[57] as 'Customer Returns'
+	   summaryMetrics.[57] as 'Customer Returns',
+	   summaryMetrics.[58] as 'Ordered Revenue'
 	   
 from 
 	 (
@@ -42,7 +44,7 @@ from
        avg(metrics.Value)
 	 for metrics.MetricTypeId
      in 
-	 ( [52], [54], [53], [55], [56], [57]) 
+	 ( [52], [54], [53], [55], [56], [57], [58]) 
 	 ) as summaryMetrics
 join td.VProduct product         on summaryMetrics.ProductId = product.Id
 join td.VSubcategory subcategory on product.SubcategoryId = subcategory.Id 
@@ -51,6 +53,6 @@ join td.VCategory category		 on subcategory.CategoryId = category.Id
                              
 join td.VParentProduct parent    on product.ParentProductId = parent.Id
 join td.VBrand brand             on product.BrandId = brand.Id 
-                             
+                             and brand.AccountId = @accId
 
 order by summaryMetrics.Date desc;
